Code design
My first major project in my master’s degree where I got a chance to
exhibit my engineering skills was the design, fabrication and coding of
a handheld Gameboy.
I did this project as a part of my Embedded System Design course which
I undertook at the University of Melbourne as a part of my Masters of
Engineering degree. This was an individual project and it took me 4
months to complete all.
The code is for realising control the game console to draw picture.